LPAL Login and 2 CPALs (#237)
This update records users, and allows two cpals to be used at the same time. The only modified file is fillLPALGUI.py.

Users input their username at the beginning, which is verified. They then input both cpals they intend to use. Whenever prompted to input a straw, the user is also given the option to input the other cpal, switching it.