Reverting to v5.21.4.
The changes to db_common files for loading old straw prep data
broke pangui pro1 -- transferring straws from LPAL to panel.

The issue is that new default args for time_in and _out were
added to the StrawPresent CTOR. In pro 1, these default args are
invoked, and we explicitly pass NULL to the DB. Insodoing, the
DB's own default args don't kick in, and the DB constraint that
time_in cannot be NULL is triggered.

The probably-good solution here is to pass timestamps to the
StrawPresent CTOR in this affected pro1 function. But need to
check everywhere that the StrawPresent CTOR is called.